* Move Obfuscate out of StringExtensions
